+++ src/Migration/2020_06_13_222012_homeworks_migration.php
@@ -16,16 +16,16 @@ class HomeworksMigration extends Migration
         Schema::create('homeworks', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name',300)->nullable();
-            $table->unsignedTinyInteger('school_id')->nullable();
-            $table->unsignedTinyInteger('divison_id')->nullable();
-            $table->unsignedTinyInteger('grade_id')->nullable();
-            $table->unsignedTinyInteger('course_id')->nullable();
-            $table->unsignedTinyInteger('teacher_id')->nullable();
+            $table->unsignedBigInteger('school_id')->nullable();
+            $table->unsignedBigInteger('divison_id')->nullable();
+            $table->unsignedBigInteger('grade_id')->nullable();
+            $table->unsignedBigInteger('course_id')->nullable();
+            $table->unsignedBigInteger('teacher_id')->nullable();
             $table->string('title',300)->nullable();
             $table->text('body')->nullable();
             $table->timestamp('due_date')->nullable();
             $table->integer('mark')->nullable();
-            $table->unsignedTinyInteger('files_id')->nullable();
+            $table->unsignedBigInteger('files_id')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
